Improvements to local restore functionality (#150)
Restoring the staging database locally can be very slow, and if you
Command+C the pgdump it will drop your local copy of the database but
not replace it with anything, resulting in you having an empty local
database.

This adds a verbose flag so you can watch pg_dump do it's work, and will
not drop the schema if there is a problem dumping the data and the dump
file is empty.

+++ scripts/rds_localrestore.sh
@@ -10,17 +10,23 @@ export PGHOST=$(aws rds describe-db-instances --filters "Name=db-instance-id,Val
 export PGPASSWORD="$(aws rds generate-db-auth-token --hostname $PGHOST --port 5432 --region $REGION --username $USERNAME)"
 
 echo "Dumping $DATABASE"
-pg_dump -Fc --no-acl --no-owner -d $DATABASE > dump.backup
+pg_dump -v -Fc --no-acl --no-owner -d $DATABASE > dump.backup
 
-export PGUSER=postgres
-export PGHOST=localhost
-export PGPASSWORD=turnout
-export PGPORT=5432
+if [ -s "dump.backup" ]
+then
+	echo "dump.backup has some data."
+    export PGUSER=postgres
+    export PGHOST=localhost
+    export PGPASSWORD=turnout
+    export PGPORT=5432
 
-echo "Dropping Schema"
-psql -d turnout -c 'DROP SCHEMA public CASCADE;CREATE SCHEMA public;'
+    echo "Dropping Schema"
+    psql -d turnout -c 'DROP SCHEMA public CASCADE;CREATE SCHEMA public;'
 
-echo "Restoring"
-pg_restore --verbose --no-owner -d turnout dump.backup
+    echo "Restoring"
+    pg_restore --verbose --no-owner -d turnout dump.backup
+else
+	echo "dump.backup is empty."
+fi
 
 rm dump.backup
